# api/docs/swagger.yaml
basePath: /
definitions:
  api.AdminInstanceResponse:
    properties:
      challenge_name:
        type: string
      container_id:
        type: string
      created_at:
        type: string
      deployment_type:
        type: string
      expires_at:
        type: string
      hosted_address:
        type: string
      instance_id:
        type: string
      port:
        type: integer
      ttl_seconds:
        type: integer
      user_id:
        type: string
      username:
        type: string
    type: object
...
  api.HTTPErrorResp:
    properties:
      error:
        example: Error occured while veifying the challenge.
        type: string
    type: object
  api.HTTPPlainResp:
    properties:
      message:
        example: Messsage in response to your request
        type: string
    type: object
...
  api.InstanceResponse:
    properties:
      challenge_name:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      hosted_address:
        type: string
      instance_id:
        type: string
      port:
        type: integer
      ttl_seconds:
        type: integer
    type: object
...
info:
  contact:
    email: contact@sdslabs.co.in
    name: SDSLabs
    url: https://chat.sdslabs.co
  description: Authenticated API for Beast CTF challenge deployment and competition
    services.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Beast API
  version: "0.2"
paths:
...
  /api/admin/instances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AdminInstanceResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all active instances
      tags:
      - admin
  /api/admin/instances/{instance_id}:
    delete:
      parameters:
      - description: Instance ID
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete an instance by ID
      tags:
      - admin
    get:
      parameters:
      - description: Instance ID
        in: path
        name: instance_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AdminInstanceResponse'
      security:
      - ApiKeyAuth: []
      summary: Get an instance by ID
      tags:
      - admin
  /api/admin/instances/challenge/{challenge_name}:
    delete:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete all instances for a challenge
      tags:
      - admin
  /api/admin/instances/user/{user_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete all instances owned by a user
      tags:
      - admin
...
  /api/instances:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.InstanceResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List the current user's instances
      tags:
      - instances
  /api/instances/{challenge_name}:
    delete:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
      security:
      - ApiKeyAuth: []
      summary: Delete the current user's challenge instance
      tags:
      - instances
    get:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Get the current user's challenge instance
      tags:
      - instances
  /api/instances/{challenge_name}/extend:
    post:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      - description: Requested extension in seconds
        in: formData
        name: seconds
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Extend the current user's challenge instance
      tags:
      - instances
  /api/instances/{challenge_name}/spawn:
    post:
      parameters:
      - description: Challenge name
        in: path
        name: challenge_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Spawn a per-user challenge instance
      tags:
      - instances
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"